Move cmodule-related config registration to tests.gof.cmodule
This untangles the `theano.configdefaults` module from the `theano.gof`
sub-package, thereby removing an import cycle.

Closes #94
